Separate drag logic from multiselection moveable.
